[OWL-384] Add exit status for commands
Every single command returns an exit code. If it is executed
properly, the return value will be 0. Some of the commands support
execution multiple modules. Once an error occurs, the return value
will be 1 and the command will be interrupted. It means that the
execution of the rest of the modules will be aborted.